workers/requirements.txt
redis==5.0.1
rq==1.16.1
gitpython==3.1.41
pyyaml==6.0.1
httpx==0.26.0
anthropic==0.18.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.3
